Indiscriminately removed global channel logic.
