+++ src/main/java/com/flightstats/hub/app/ClusterHubBindings.java
@@ -29,10 +29,6 @@ class ClusterHubBindings extends AbstractModule {
     protected void configure() {
         logger.info("starting server {} ", HubHost.getLocalName());
 
-        bind(ChannelService.class).to(ChannelNameService.class).asEagerSingleton();
-        bind(ChannelService.class)
-                .annotatedWith(Names.named(ChannelService.DELEGATE))
-                .to(GlobalChannelService.class).asEagerSingleton();
         bind(AwsConnectorFactory.class).asEagerSingleton();
         bind(S3Config.class).asEagerSingleton();
         bind(ContentService.class)
